feat: edit annotation queue
